Rx swift

+++ FootPrintGram/Util/FireBaseUtil.swift
@@ -12,92 +12,26 @@ import FirebaseDatabase
 import FirebaseAuth
 import MapKit
 import PromiseKit
-
+import RxSwift
 
 class FireBaseUtil {
     
     static let shared = FireBaseUtil()
     
-    var userInfo = UserInfo.shared
+    private var userInfo = UserInfo.shared
     
-    let settingMeta = StorageMetadata.init()
-    let storage = Storage.storage().reference().child("userPostImage")
+    private let settingMeta = StorageMetadata.init()
+    private let storage = Storage.storage().reference().child("userPostImage")
     
-    var mainData: FootPrintAnnotationList!
-    var ref = Database.database().reference()
-    
-    var lastData: NSDictionary!
+    private var ref = Database.database().reference();
     
     var recentStorage: StorageReference!
     
-    func callAllPostWithLoadingUI() -> Promise<String> {
-        return Promise<String> { [unowned self] scene -> Void in
-            self.ref.child("footprintPosts").child(self.userInfo.uid).observeSingleEvent(of: .value, with: { (snapshot) in
-                if(!snapshot.exists()) {
-                    scene.fulfill("NotExist")
-                }
-                guard let value = (snapshot.value as? NSDictionary) else { return }
-                
-                if(value != self.lastData) {
-                    self.lastData = value
-                    FootPrintAnnotationList.shared.fpaList?.removeAll()
-                    for child in snapshot.children {
-                        let snap = child as! DataSnapshot
-                        let key = snap.key
-                        let value = snap.value as! NSDictionary
-                        print(key, value)
-                        let cood = CLLocationCoordinate2D.init(latitude: Double(value["latitude"] as! String)! as CLLocationDegrees, longitude: Double(value["longitude"] as! String)! as CLLocationDegrees)
-                        let title = value["name"] as? String
-                        let imageUrl = value["profileImageURL"] as? String
-                        let created = value["created"] as? String
-                        let item = FootPrintAnnotation.init(coordinate: cood, title: title!, imageUrl: imageUrl!, created: created, id: key)
-                        //                        item.id = key
-                        FootPrintAnnotationList.shared.fpaList!.append(item)
-                    }
-                    self.mainData = FootPrintAnnotationList.shared
-                    scene.fulfill("SUC")
-                }
-            }, withCancel: nil)
-        }
-    }
-    
-    private func callPostUpdateWithLoadingUI(data: [String:String], url: String) -> Promise<String> {
-        var temp_data = data
-        temp_data["profileImageURL"] = url
-        let r_data = (temp_data as NSDictionary)
-        return Promise<String> { scene -> Void in
-            Database.database().reference().child("footprintPosts").child(self.userInfo.uid).childByAutoId().setValue(r_data, withCompletionBlock: { (err, ref) in
-                guard let error = err else { return scene.fulfill("SUC") }
-                return scene.reject(error)
-            })
-        }
-    }
-    
-    private func callPhotoPut(_ title: String, _ data: Data) -> Promise<String> {
-        settingMeta.contentType = "image/jepg"
-        let r_storage = storage.child(userInfo.uid).child(title + Date.init().description)
-        recentStorage = r_storage
-        return Promise<String> { scene -> Void in
-            r_storage.putData(data, metadata: settingMeta, completion: { (ref, err) in
-                guard let err = err else { return scene.fulfill("SUC") }
-                return scene.reject(err)
-            })
-        }
-    }
-    
-    private func callPhotoDownUrl() -> Promise<String> {
-        return Promise<String> { scene -> Void in
-            guard let rStorage = recentStorage else { return scene.reject(tempError.one) }
-            rStorage.downloadURL(completion: { (url, err) in
-                guard let err = err else { return scene.fulfill(url!.absoluteString) }
-                return scene.reject(err)
-            })
-        }
-    }
-    
     private func callDeleteById(_ ind: Int, _ id: String) -> Promise<String> {
         return Promise<String> { scene -> Void in
-            let deleteStorage = Storage.storage().reference(forURL: self.mainData.fpaList![ind].post.imageUrl!)
+
+            guard let imgUrl = FootPrintAnnotationList.shared.itemImageUrl(ind: ind) else { return }
+            let deleteStorage = Storage.storage().reference(forURL: imgUrl)
             deleteStorage.delete(completion: { (err) in
                 guard let errorCode = (err as NSError?)?.code else { return scene.fulfill("SUC") }
                 guard let error = StorageErrorCode(rawValue: errorCode) else { return }
@@ -114,19 +48,11 @@ class FireBaseUtil {
     }
     
     private func callDeleteDBById(_ id: String) -> Promise<String> {
-        return Promise<String> { scene in self.ref.child("footprintPosts").child(self.userInfo.uid).child(id).removeValue(completionBlock: { (err, ref) in
+        return Promise<String> { scene in self.ref.child("footprintPosts").child(self.userInfo.o_uid).child(id).removeValue(completionBlock: { (err, ref) in
                 guard let err = err else { return scene.fulfill("SUC") }
                 return scene.reject(err)
             })
         }
-    }
-    
-    func totalFunc(_ title: String, _ data: Data, data2: [String:String]) -> Promise<String> {
-        return callPhotoPut(title, data).then {
-                _ in self.callPhotoDownUrl()
-            }.then {
-                url in self.callPostUpdateWithLoadingUI(data: data2, url: url)
-            }
     }
     
     func deleteTotalFunc(_ ind: Int,_ id: String) -> Promise<String> {
@@ -136,9 +62,132 @@ class FireBaseUtil {
     }
     
     enum tempError: Error { // 전역화
-        case one
-        case two
-        case three
+        case NotExist
+        case CastingError
+        case UnwrappedError
+    }
+    
+}
+
+extension FireBaseUtil {
+    
+    func fbAllPostLoad(completed: @escaping (DataSnapshot?) -> Void, cancel: ((Error) -> Void)?) {
+        DispatchQueue.global().async {
+            self.ref.child("footprintPosts").child(self.userInfo.o_uid).observeSingleEvent(of: .value, with: completed, withCancel: cancel)
+        }
+    }
+    
+    func rxAllPostLoad() -> Observable<String> {
+        return Observable.create { seal in
+            self.fbAllPostLoad(completed: { (snapshot) in
+                guard let snapshot = snapshot else { return seal.onError(tempError.UnwrappedError) }
+                
+                if(!snapshot.exists()) {
+                    seal.onError(tempError.NotExist)
+                }
+                FootPrintAnnotationList.shared.removeAllItem()
+                
+                for child in snapshot.children {
+                    let snap = child as! DataSnapshot
+                    let key = snap.key
+                    let value = snap.value as! NSDictionary
+                    print(key, value)
+                    let cood = CLLocationCoordinate2D.init(latitude: Double(value["latitude"] as! String)! as CLLocationDegrees, longitude: Double(value["longitude"] as! String)! as CLLocationDegrees)
+                    let title = value["name"] as? String
+                    let imageUrl = value["profileImageURL"] as? String
+                    let created = value["created"] as? String
+                    let item = FootPrintAnnotation.init(coordinate: cood, title: title!, imageUrl: imageUrl!, created: created, id: key)
+                    FootPrintAnnotationList.shared.appendItem(item: item)
+                }
+                seal.onNext("SUC")
+            }, cancel: nil)
+            return Disposables.create()
+        }
+    }
+    
+    func rxAllPostLoad2() -> Observable<[FootPrintAnnotation]> {
+        return Observable.create { seal in
+            self.fbAllPostLoad(completed: { (snapshot) in
+                guard let snapshot = snapshot else { return seal.onError(tempError.UnwrappedError) }
+                
+                if(!snapshot.exists()) {
+                    seal.onError(tempError.NotExist)
+                }
+                FootPrintAnnotationList.shared.removeAllItem()
+                
+                for child in snapshot.children {
+                    let snap = child as! DataSnapshot
+                    let key = snap.key
+                    let value = snap.value as! NSDictionary
+                    print(key, value)
+                    let cood = CLLocationCoordinate2D.init(latitude: Double(value["latitude"] as! String)! as CLLocationDegrees, longitude: Double(value["longitude"] as! String)! as CLLocationDegrees)
+                    let title = value["name"] as? String
+                    let imageUrl = value["profileImageURL"] as? String
+                    let created = value["created"] as? String
+                    let item = FootPrintAnnotation.init(coordinate: cood, title: title!, imageUrl: imageUrl!, created: created, id: key)
+                    FootPrintAnnotationList.shared.appendItem(item: item)
+                }
+                seal.onNext(FootPrintAnnotationList.shared.getList()!)
+            }, cancel: nil)
+            return Disposables.create()
+        }
+    }
+    
+    func fbUploadData(data: [String:String], url: String, completed: @escaping (Error?, DatabaseReference?) -> Void) {
+        var temp_data = data
+        temp_data["profileImageURL"] = url
+        let r_data = (temp_data as NSDictionary)
+        DispatchQueue.global().async {
+            Database.database().reference().child("footprintPosts").child(self.userInfo.o_uid).childByAutoId().setValue(r_data, withCompletionBlock: completed)
+        }
+    }
+
+    func rxUploadData(data: [String:String], url: String) -> Observable<Bool?> {
+        return Observable.create { seal in
+            self.fbUploadData(data: data, url: url) { (err, ref) in
+                guard let ref = ref else { return seal.onError(err!) }
+                seal.onNext(true)
+                seal.onCompleted()
+            }
+            return Disposables.create()
+        }
+    }
+
+    func fbUploadImage(_ title: String, _ data: Data, completed: @escaping (StorageMetadata?, Error?) -> Void) {
+        settingMeta.contentType = "image/jepg"
+        let r_storage = storage.child(userInfo.o_uid).child(title + Date.init().description)
+        recentStorage = r_storage
+        
+        DispatchQueue.global().async {
+            r_storage.putData(data, metadata: self.settingMeta, completion: completed)
+        }
+    }
+
+    func rxUploadImage(_ title: String, _ data: Data) -> Observable<Bool?> {
+        return Observable.create { seal in
+            self.fbUploadImage(title, data) { smeta, err in
+                guard let smeta = smeta else { return seal.onError(err!) }
+                seal.onNext(true)
+            }
+            return Disposables.create()
+        }
+    }
+
+    func fbGetImageUrl(completed: @escaping (URL?, Error?) -> Void) {
+        guard let rStorage = recentStorage else { return }
+        DispatchQueue.global().async {
+            rStorage.downloadURL(completion: completed)
+        }
+    }
+
+    func rxGetImageUrl() -> Observable<String?> {
+        return Observable.create { seal in
+            self.fbGetImageUrl(){ url, err in
+                guard let url = url else { return seal.onError(err!) }
+                seal.onNext(url.absoluteString)
+            }
+            return Disposables.create()
+        }
     }
     
 }
